2.0.0-beta2 (#16)
* Use short array syntax for color set assignment

Replaced the `list()` syntax with short array syntax `[]` for assigning `$darkColor` and `$lightColor`. This improves code readability and aligns with modern PHP standards.

* Add release configuration for automated changelogs

This commit introduces a `.github/release.yml` file to automate the generation of changelogs during releases. It categorizes changes into breaking changes, new features, bug fixes, and other updates based on labels. This improves clarity and streamlines the release process.

* Update PHP and dependencies to latest versions

Upgraded the PHP requirement to 8.2 and updated development dependencies, including PHPUnit to 11.5 and Testbench to 9.7. These changes ensure compatibility with newer features and maintain alignment with the latest ecosystem standards.

* Remove PHP 8.1 from GitHub Actions workflow.

This update streamlines the supported PHP versions in the CI pipeline by dropping PHP 8.1. Only PHP 8.2 and newer are now tested, reflecting a focus on more recent versions.

* Use DataProvider attribute in Avatar tests

Replaced traditional `@dataProvider` annotations with `#[DataProvider]` attributes in test methods. This modernizes the code and improves readability while maintaining functionality. Additionally, updated provider methods to be `static` to match the attribute requirements.

* Update type hints and remove @expectedException annotations

Added the `mixed` return type hint to the `invokeMethod` function for stricter typing. Removed redundant `@expectedException` annotations as they are no longer needed with modern PHPUnit exception handling.

* Make test data providers static and use attributes for tests

Updated all data provider methods to be static for compatibility with PHPUnit's attribute-based DataProvider system. Replaced legacy `@dataProvider` annotations with modern `#[DataProvider]` attributes for improved readability and maintainability.

* Implement attributes-based DataProvider in LarvatarTraitTest

Replaced traditional docblock-based `@dataProvider` annotations with PHPUnit's native attributes. This improves readability and aligns with modern PHPUnit versions. Additionally, some methods were refactored for better consistency and clarity.

* Refactor formatting and inline concatenations

Reformat docblock parameters for consistent style and improve readability. Replace inline string concatenations with clearer spacing and streamline conditional logic.

* Refactor docblocks and reorganize `getHexColor` method.

Standardize docblock formatting for consistency and clarity. Reorder the `getHexColor` method for better code structure and readability.

* Simplify and correct docblock annotations.

Updated the docblock formatting for better consistency and readability. Added type specificity to the return annotation to improve clarity and type safety.

* Update PHPDoc return types for matrix methods

Clarified the return types in PHPDoc annotations to provide more precise type information for symmetric matrix and matrix generation methods. This improves code readability and developer understanding of method outputs.

* Cast lightness values to float in Avatar setters

Ensure `backgroundLightness` and `textLightness` values are explicitly cast to float after clamping. This improves type clarity and ensures consistent data handling across the application.

* Update bug report template with new version options

Added support for versions 2.0.x and 1.4.x in the bug report dropdown menu. This ensures users can accurately select their software version when submitting issues.

+++ src/Avatar.php
@@ -138,7 +138,7 @@ abstract class Avatar
      */
     public function setBackgroundLightness(float $backgroundLightness): void
     {
-        $this->backgroundLightness = clamp($backgroundLightness, 0, 1);
+        $this->backgroundLightness = (float)clamp($backgroundLightness, 0, 1);
     }
 
     /**
@@ -159,7 +159,7 @@ abstract class Avatar
      */
     public function setTextLightness(float $textLightness): void
     {
-        $this->textLightness = clamp($textLightness, 0, 1);
+        $this->textLightness = (float)clamp($textLightness, 0, 1);
     }
 
     /**
